+++ CMakeLists.txt
@@ -12,6 +12,7 @@ find_package(catkin REQUIRED COMPONENTS
   rospy
   std_msgs
   message_filters
+  message_generation
 )
 
 find_package(PCL 1.2 REQUIRED)
@@ -55,18 +56,16 @@ add_definitions(${PCL_DEFINITIONS})
 ##   * add every package in MSG_DEP_SET to generate_messages(DEPENDENCIES ...)
 
 ## Generate messages in the 'msg' folder
-# add_message_files(
-#   FILES
-#   Message1.msg
-#   Message2.msg
-# )
+add_message_files(
+  FILES 
+  ArmInputs.msg
+)
 
 ## Generate services in the 'srv' folder
-# add_service_files(
-#   FILES
-#   Service1.srv
-#   Service2.srv
-# )
+add_service_files(
+  FILES
+  Corrections.srv
+)
 
 ## Generate actions in the 'action' folder
 # add_action_files(
@@ -76,10 +75,10 @@ add_definitions(${PCL_DEFINITIONS})
 # )
 
 ## Generate added messages and services with any dependencies listed here
-# generate_messages(
-#   DEPENDENCIES
-#   std_msgs
-# )
+generate_messages(
+  DEPENDENCIES
+  std_msgs
+)
 
 ################################################
 ## Declare ROS dynamic reconfigure parameters ##
@@ -113,7 +112,7 @@ add_definitions(${PCL_DEFINITIONS})
 catkin_package(
     #INCLUDE_DIRS include
     #LIBRARIES Rover
-    CATKIN_DEPENDS roscpp rospy std_msgs message_filters
+    CATKIN_DEPENDS roscpp rospy std_msgs message_filters message_runtime
     DEPENDS system_lib
 )
