+++ package.xml
@@ -29,6 +29,9 @@
   <exec_depend>ros_phoenix</exec_depend>
   <exec_depend>message_filters</exec_depend>
 
+  <build_depend>message_generation</build_depend>
+  <exec_depend>message_runtime</exec_depend>
+
 
   <!-- The export tag contains other, unspecified, tags -->
   <export>
