+++ srv/Corrections.srv
@@ -1,0 +1,3 @@
+float32 correction1
+---
+bool updated
